refactor: inherit proxy unleash from base class and tap into existing DTOs for variants
